requirements_new.txt
requests==2.31.0
langchain_core==0.1.42
langchain_groq==0.1.2
langchain_openai==0.1.3
langchain==0.1.16
deepgram-sdk==3.2.5
pyaudio==0.2.14
djangorestframework==3.15.1
django-cors-headers==4.3.1
pyjwt==2.8.0
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.1





requests==2.31.0
langchain_core==0.1.42
langchain_groq==0.1.2
langchain_openai==0.1.3
langchain==0.1.16
deepgram-sdk==3.2.5
#pyaudio==0.2.14
djangorestframework==3.15.1
django-cors-headers==4.3.1
gunicorn